[catnap] Enhancement: Add socket
